雪花算法改良，最后12个序列号支持不同毫秒递增，这样生成的id才能支持后面的rang范围分片算法查询

+++ src/main/java/com/plat/db/algorithm/InileRangAlgorithm.java
@@ -13,7 +13,7 @@ import io.shardingsphere.api.algorithm.sharding.RangeShardingValue;
 import io.shardingsphere.api.algorithm.sharding.standard.RangeShardingAlgorithm;
 import io.shardingsphere.core.util.InlineExpressionParser;
 
-public final class InileRangAlgorithm implements RangeShardingAlgorithm<Integer>{
+public final class InileRangAlgorithm implements RangeShardingAlgorithm<Long>{
 
 	private Logger logger = LoggerFactory.getLogger(InileRangAlgorithm.class);
 
@@ -26,7 +26,7 @@ public final class InileRangAlgorithm implements RangeShardingAlgorithm<Integer>
 	private final Closure<?> closure;
 
 	@Override
-	public Collection<String> doSharding(Collection<String> availableTargetNames, RangeShardingValue<Integer> shardingValue) {
+	public Collection<String> doSharding(Collection<String> availableTargetNames, RangeShardingValue<Long> shardingValue) {
 				
         Set<String> result = new LinkedHashSet<>();
